easi/tasks/registry.py
"""Task registry with auto-discovery from task.yaml files.

Scans easi/tasks/*/task.yaml to discover available tasks (benchmarks).

Usage:
    list_tasks() → ["dummy_task", "objectnav_hm3d", ...]
    get_task_entry("dummy_task") → TaskEntry(...)
    load_task_class("dummy_task") → DummyTask class
"""

from __future__ import annotations

import importlib
import logging
from dataclasses import dataclass
from pathlib import Path

import yaml

logger = logging.getLogger("easi.tasks.registry")


@dataclass
class TaskEntry:
    """Registry entry for a task."""

    name: str
    display_name: str
    description: str
    simulator_key: str
    task_class: str  # fully qualified class name
    action_space: list[str]
    max_steps: int
    config_path: Path


# Module-level registry populated on first access
_registry: dict[str, TaskEntry] | None = None


def _discover_tasks() -> dict[str, TaskEntry]:
    """Scan task directories for task.yaml files."""
    tasks_dir = Path(__file__).parent
    entries: dict[str, TaskEntry] = {}

    for task_yaml_path in sorted(tasks_dir.glob("*/task.yaml")):
        try:
            config = yaml.safe_load(task_yaml_path.read_text())
        except Exception as e:
            logger.warning("Failed to load %s: %s", task_yaml_path, e)
            continue

        name = config["name"]
        entries[name] = TaskEntry(
            name=name,
            display_name=config.get("display_name", name),
            description=config.get("description", ""),
            simulator_key=config["simulator"],
            task_class=config["task_class"],
            action_space=config.get("action_space", []),
            max_steps=config.get("max_steps", 500),
            config_path=task_yaml_path,
        )

        logger.debug("Discovered task: %s (simulator: %s)", name, config["simulator"])

    return entries


def _get_registry() -> dict[str, TaskEntry]:
    """Get the task registry, discovering on first access."""
    global _registry
    if _registry is None:
        _registry = _discover_tasks()
    return _registry


def get_task_entry(name: str) -> TaskEntry:
    """Look up a task entry by name.

    Raises:
        KeyError: If the task is not found.
    """
    registry = _get_registry()
    if name not in registry:
        available = list_tasks()
        raise KeyError(f"Task '{name}' not found. Available: {available}")
    return registry[name]


def list_tasks() -> list[str]:
    """List all registered task names."""
    return sorted(_get_registry().keys())


def load_task_class(name: str):
    """Import and return the task class for the given name."""
    entry = get_task_entry(name)
    return _import_class(entry.task_class)


def _import_class(fully_qualified_name: str):
    """Import a class from its fully qualified name."""
    module_path, class_name = fully_qualified_name.rsplit(".", 1)
    module = importlib.import_module(module_path)
    return getattr(module, class_name)


def refresh() -> None:
    """Force re-discovery of tasks."""
    global _registry
    _registry = None
